+++ cfspopcon/__init__.py
@@ -11,6 +11,7 @@ from .deprecation_handler import handle_deprecated_arguments
 from .formulas.atomic_data import AtomicData
 from .input_file_handling import process_input_dictionary, read_case
 from .plotting import read_plot_style
+from .plugins import PluginClashError, PluginReport, register_plugin
 from .unit_handling import (
     convert_to_default_units,
     convert_units,
@@ -23,6 +24,8 @@ __all__ = [
     "Algorithm",
     "AtomicData",
     "CompositeAlgorithm",
+    "PluginClashError",
+    "PluginReport",
     "convert_to_default_units",
     "convert_units",
     "file_io",
@@ -33,6 +36,7 @@ __all__ = [
     "process_input_dictionary",
     "read_case",
     "read_plot_style",
+    "register_plugin",
     "set_default_units",
     "shaping_and_selection",
 ]